refactor: 디렉터리 정리 및  이름 변경
