+++ src/es/Store/Interface.ts
