+++ tsconfig.json
@@ -5,10 +5,10 @@
     "esModuleInterop": true,
     "sourceMap": true,
 		"paths": {
-			"@Domain/*": ["./src/js/domain/*"],
-			"@Display/*": ["./src/js/display/*"],
-			"@Utils/*": ["./src/js/utils/*"],
-			"@Constants/*": ["./src/js/constants/*"],
+			"@Store/*": ["./src/es/Store/*"],
+			"@Display/*": ["./src/es/display/*"],
+			"@Utils/*": ["./src/es/utils/*"],
+			"@Constants/*": ["./src/es/constants/*"],
 		}
   }
 }
