+++ webpack.config.js
@@ -10,10 +10,10 @@ module.exports = {
     extensions: ['.js', '.ts', '.css', '.scss'],
     alias: {
       '@Styles': path.resolve(__dirname, 'src/styles'),
-      '@Display': path.resolve(__dirname, 'src/js/display'),
-      '@Domain': path.resolve(__dirname, 'src/js/domain'),
-      '@Utils': path.resolve(__dirname, 'src/js/utils'),
-      '@Constants': path.resolve(__dirname, 'src/js/constants'),
+      '@Display': path.resolve(__dirname, 'src/es/display'),
+      '@Store': path.resolve(__dirname, 'src/es/Store'),
+      '@Utils': path.resolve(__dirname, 'src/es/utils'),
+      '@Constants': path.resolve(__dirname, 'src/es/constants'),
     },
   },
   devServer: {
